fixed index js files and renamed files
