+++ package.json
@@ -3,34 +3,29 @@
   "version": "1.0.0",
   "description": "Google Books Search",
   "main": "server.js",
-  "proxy": "http://localhost:3001",
+  "proxy": "http://localhost:3002",
   "scripts": {
     "start": "if-env NODE_ENV=production && npm run start:prod || npm run start:dev",
-    "start:prod": "node server.js",
-    "start:dev": "concurrently \"nodemon --ignore 'client/*'\" \"npm run client\"",
-    "client": "cd client && npm run start",
-    "seed": "node scripts/seedDB.js",
-    "install": "cd client && npm install",
-    "build": "cd client && npm run build",
-    "heroku-postbuild": "npm run build"
+   "start:prod": "node server.js",
+   "start:dev": "concurrently \"nodemon --ignore 'client/*'\" \"npm run client\"",
+   "client": "cd client && npm run start",
+   "install": "cd client && npm install",
+   "build": "cd client && npm run build",
+   "heroku-postbuild": "npm run build"
   },
   "dependencies": {
     "axios": "^0.19.0",
     "body-parser": "^1.19.0",
     "bootstrap": "^4.3.1",
+    "concurrently": "^4.1.1",
     "express": "^4.17.1",
     "if-env": "^1.0.4",
     "mongoose": "^5.6.4",
+    "nodemon": "^1.19.1",
     "react": "^16.8.6",
     "react-bootstrap": "^1.0.0-beta.9",
     "react-dom": "^16.8.6",
     "react-scripts": "3.0.1"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
   },
   "eslintConfig": {
     "extends": "react-app"
